add something with changing item excel

+++ Excel2Lua/DataStruct/AB_CumulativeRecharge.cs
@@ -42,8 +42,8 @@ namespace Excel2Lua
 			bool bReturn = true;
 			foreach (CumulativeRechargeInfo info in m_listOwn)
 			{
-				bool bCheckMale = ItemMgr.Instance.CheckItemAndLogError(m_strExcelSheet, (ushort)info.malereward_itemid, info.malereward_itemcount, info.malereward_itemvalidity);
-				bool bCheckFemale = ItemMgr.Instance.CheckItemAndLogError(m_strExcelSheet, (ushort)info.femalereward_itemid, info.femalereward_itemcount, info.femalereward_itemvalidity);
+				bool bCheckMale = ItemMgr.Instance.CheckItemAndLogError(m_strExcelSheet, info.malereward_itemid, info.malereward_itemcount, info.malereward_itemvalidity);
+				bool bCheckFemale = ItemMgr.Instance.CheckItemAndLogError(m_strExcelSheet, info.femalereward_itemid, info.femalereward_itemcount, info.femalereward_itemvalidity);
 				if (!bCheckFemale || !bCheckMale)
 				{
 					bReturn = false;
@@ -219,7 +219,7 @@ namespace Excel2Lua
 			[ExcelHeader("累计数量")]
 			public string requirenum { get; set; }
 			[ExcelHeader("男奖励物品编号")]
-			public int malereward_itemid { get; set; }
+			public uint malereward_itemid { get; set; }
 			[ExcelHeader("男奖励物品名称")]
 			public string malereward_itemname { get; set; }
 			[ExcelHeader("男奖励物品数量")]
@@ -227,7 +227,7 @@ namespace Excel2Lua
 			[ExcelHeader("男奖励物品时限")]
 			public int malereward_itemvalidity { get; set; }
 			[ExcelHeader("女奖励物品编号")]
-			public int femalereward_itemid { get; set; }
+			public uint femalereward_itemid { get; set; }
 			[ExcelHeader("女奖励物品名称")]
 			public string femalereward_itemname { get; set; }
 			[ExcelHeader("女奖励物品数量")]
@@ -235,9 +235,9 @@ namespace Excel2Lua
 			[ExcelHeader("女奖励物品时限")]
 			public int femalereward_itemvalidity { get; set; }
 			[ExcelHeader("金券奖励")]
-			public int money { get; set; }
+			public uint money { get; set; }
 			[ExcelHeader("绑定M币奖励")]
-			public int bindmcoin { get; set; }
+			public uint bindmcoin { get; set; }
 
 			public bool IsSameKind(CumulativeRechargeInfo info)
 			{
@@ -294,8 +294,8 @@ namespace Excel2Lua
 				}
 			}
 
-			public void AddItem(int nMaleItemID, int nMaleItemCount, int nMaleItemValidity
-				, int nFemaleItemID, int nFemaleItemCount, int nFemaleItemValidity)
+			public void AddItem(uint nMaleItemID, int nMaleItemCount, int nMaleItemValidity
+				, uint nFemaleItemID, int nFemaleItemCount, int nFemaleItemValidity)
 			{
 				AddItem(true, new Lua_Item(nMaleItemID, nMaleItemCount, nMaleItemValidity));
 				AddItem(false, new Lua_Item(nFemaleItemID, nFemaleItemCount, nFemaleItemValidity));
