+++ Excel2Lua/DataStruct/AB_FresherActivity.cs
@@ -42,8 +42,8 @@ namespace Excel2Lua
 			bool bReturn = true;
 			foreach (FresherActivityInfo info in m_listOwn)
 			{
-				bool bCheckMale = ItemMgr.Instance.CheckItemAndLogError(m_strExcelSheet, (ushort)info.malereward_itemid, info.malereward_itemcount, info.malereward_itemvalidity);
-				bool bCheckFemale = ItemMgr.Instance.CheckItemAndLogError(m_strExcelSheet, (ushort)info.femalereward_itemid, info.femalereward_itemcount, info.femalereward_itemvalidity);
+				bool bCheckMale = ItemMgr.Instance.CheckItemAndLogError(m_strExcelSheet, info.malereward_itemid, info.malereward_itemcount, info.malereward_itemvalidity);
+				bool bCheckFemale = ItemMgr.Instance.CheckItemAndLogError(m_strExcelSheet, info.femalereward_itemid, info.femalereward_itemcount, info.femalereward_itemvalidity);
 				if (!bCheckFemale || !bCheckMale)
 				{
 					bReturn = false;
@@ -238,7 +238,7 @@ namespace Excel2Lua
 			[ExcelHeader("天数")]
 			public int day { get; set; }
 			[ExcelHeader("男奖励物品编号")]
-			public int malereward_itemid { get; set; }
+			public uint malereward_itemid { get; set; }
 			[ExcelHeader("男奖励物品名称")]
 			public string malereward_itemname { get; set; }
 			[ExcelHeader("男奖励物品数量")]
@@ -246,7 +246,7 @@ namespace Excel2Lua
 			[ExcelHeader("男奖励物品时限")]
 			public int malereward_itemvalidity { get; set; }
 			[ExcelHeader("女奖励物品编号")]
-			public int femalereward_itemid { get; set; }
+			public uint femalereward_itemid { get; set; }
 			[ExcelHeader("女奖励物品名称")]
 			public string femalereward_itemname { get; set; }
 			[ExcelHeader("女奖励物品数量")]
@@ -254,11 +254,11 @@ namespace Excel2Lua
 			[ExcelHeader("女奖励物品时限")]
 			public int femalereward_itemvalidity { get; set; }
 			[ExcelHeader("金券奖励")]
-			public int money { get; set; }
+			public uint money { get; set; }
 			[ExcelHeader("绑定M币奖励")]
-			public int bindmcoin { get; set; }
+			public uint bindmcoin { get; set; }
 			[ExcelHeader("VIP男奖励物品编号")]
-			public int vipmalereward_itemid { get; set; }
+			public uint vipmalereward_itemid { get; set; }
 			[ExcelHeader("VIP男奖励物品名称")]
 			public string vipmalereward_itemname { get; set; }
 			[ExcelHeader("VIP男奖励物品数量")]
@@ -266,7 +266,7 @@ namespace Excel2Lua
 			[ExcelHeader("VIP男奖励物品时限")]
 			public int vipmalereward_itemvalidity { get; set; }
 			[ExcelHeader("VIP女奖励物品编号")]
-			public int vipfemalereward_itemid { get; set; }
+			public uint vipfemalereward_itemid { get; set; }
 			[ExcelHeader("VIP女奖励物品名称")]
 			public string vipfemalereward_itemname { get; set; }
 			[ExcelHeader("VIP女奖励物品数量")]
@@ -274,9 +274,9 @@ namespace Excel2Lua
 			[ExcelHeader("VIP女奖励物品时限")]
 			public int vipfemalereward_itemvalidity { get; set; }
 			[ExcelHeader("VIP金券奖励")]
-			public int vipmoney { get; set; }
+			public uint vipmoney { get; set; }
 			[ExcelHeader("VIP绑定M币奖励")]
-			public int vipbindmcoin { get; set; }
+			public uint vipbindmcoin { get; set; }
 
 			public bool IsSameKind(FresherActivityInfo info)
 			{
@@ -322,11 +322,11 @@ namespace Excel2Lua
 		{
 			int m_nIndex { get; set; }
 			string m_strOnlineTime { get; set; }
-			int m_nMoney { get; set; }
+			uint m_nMoney { get; set; }
 			public List<Lua_Item> m_listMaleItem;
 			public List<Lua_Item> m_listFemaleItem;
 
-			public Lua_Online(int nIndex, string strOnlineTime, int nMoney)
+			public Lua_Online(int nIndex, string strOnlineTime, uint nMoney)
 			{
 				m_nIndex = nIndex;
 				m_strOnlineTime = strOnlineTime;
@@ -347,8 +347,8 @@ namespace Excel2Lua
 				}
 			}
 
-			public void AddItem(int nMaleItemID, int nMaleItemCount, int nMaleItemValidity
-				, int nFemaleItemID, int nFemaleItemCount, int nFemaleItemValidity)
+			public void AddItem(uint nMaleItemID, int nMaleItemCount, int nMaleItemValidity
+				, uint nFemaleItemID, int nFemaleItemCount, int nFemaleItemValidity)
 			{
 				AddItem(true, new Lua_Item(nMaleItemID, nMaleItemCount, nMaleItemValidity));
 				AddItem(false, new Lua_Item(nFemaleItemID, nFemaleItemCount, nFemaleItemValidity));
