+++ Excel2Lua/Processor/E2L_Box.cs
@@ -89,7 +89,7 @@ namespace Excel2Lua
 		class BoxInfo
 		{
 			[ExcelHeader("男奖励物品编号")]
-			public int malereward_itemid { get; set; }
+			public uint malereward_itemid { get; set; }
 			[ExcelHeader("男奖励物品名称")]
 			public string malereward_itemname { get; set; }
 			[ExcelHeader("男奖励物品数量")]
@@ -99,7 +99,7 @@ namespace Excel2Lua
 			[ExcelHeader("男奖励权重")]
 			public float malereward_weight { get; set; }
 			[ExcelHeader("女奖励物品编号")]
-			public int femalereward_itemid { get; set; }
+			public uint femalereward_itemid { get; set; }
 			[ExcelHeader("女奖励物品名称")]
 			public string femalereward_itemname { get; set; }
 			[ExcelHeader("女奖励物品数量")]
@@ -257,8 +257,8 @@ namespace Excel2Lua
 				bool bReturn = true;
 				foreach (BoxInfo info in m_listReward)
 				{
-					bool bCheckMale = ItemMgr.Instance.CheckItemAndLogError(CustomDefine.BOX_EXCEL_NAME, (ushort)info.malereward_itemid, info.malereward_itemcount, info.malereward_itemvalidity);
-					bool bCheckFemale = ItemMgr.Instance.CheckItemAndLogError(CustomDefine.BOX_EXCEL_NAME, (ushort)info.femalereward_itemid, info.femalereward_itemcount, info.femalereward_itemvalidity);
+					bool bCheckMale = ItemMgr.Instance.CheckItemAndLogError(CustomDefine.BOX_EXCEL_NAME, info.malereward_itemid, info.malereward_itemcount, info.malereward_itemvalidity);
+					bool bCheckFemale = ItemMgr.Instance.CheckItemAndLogError(CustomDefine.BOX_EXCEL_NAME, info.femalereward_itemid, info.femalereward_itemcount, info.femalereward_itemvalidity);
 					if (!bCheckFemale || !bCheckMale)
 					{
 						bReturn = false;
